Fix RL/SFT correctness bugs found in paper-spec review
- GRPO: trim rollouts at first EOS — batched generate() pads finished
  samples with eos and the loss was training on those positions
- Wire up --resume-from-lora (was parsed but never used; the v4-long
  "resume" silently restarted from AV-SFT). KL ref stays AV-SFT init;
  co-trained critic now saved per-iter and resumed from critic/ subdir
- vLLM rollouts: SteeringVector must be 3-D [1,1,d] — 2-D takes
  vllm-lens's broadcast branch and steers EVERY token, ignoring
  position_indices
- FVE: use the paper's raw-variance baseline (≈0.72), not the inflating
  meannorm baseline (≈0.94); both logged, old curves not comparable
- AR critic: strip final RMSNorm in init_critic_from_base (matches
  NLACriticModel.from_pretrained + design §4); recorded as
  final_norm_stripped in ar_meta.json, old ckpts default to norm-kept
- Tokenizer from --base-ckpt/--av-ckpt instead of hardcoded Qwen3-8B
- AR-SFT: add_special_tokens=False to match RL critic scoring; skip
  (not truncate) over-length critic prompts to protect suffix anchor
- NaN guard before optim.step (was after — Adam moments corruptible)
- Asserts: temperature==1.0 (raw-logits old_logp), max_rows<=eval_skip_rows
  (training pool contained literal eval rows), vLLM logprob lookups
- vLLM trainer: doc-disjoint eval filter ported from twin; 4.57.1
  BatchEncoding compat in schema; value_head dtype align under QLoRA

+++ nla/models.py
@@ -204,7 +204,13 @@ class NLACriticModel(PreTrainedModel):
         # out.last_hidden_state: [B, T, d_model] — post last-kept transformer block,
         # pre-final-LN (norm was replaced with Identity in from_pretrained).
         h = out.last_hidden_state
-        return NLACriticOutput(values=self.value_head(h), backbone_last_hidden=h)
+        # Align dtype before the value_head matmul: under 4-bit QLoRA,
+        # prepare_model_for_kbit_training casts the backbone's final norm to
+        # fp32, so h may be fp32 while value_head is bf16 (or vice-versa).
+        return NLACriticOutput(
+            values=self.value_head(h.to(self.value_head.weight.dtype)),
+            backbone_last_hidden=h,
+        )
 
     def get_input_embeddings(self):
         return self.backbone.get_input_embeddings()
